+++ .github/workflows/ci.yml
@@ -77,45 +77,11 @@ jobs:
       - name: Lint
         run: |
           cargo clippy --all-targets --release --locked -- -D clippy::all
-          target/release/deno_lint benchmarks/benchmarks.ts
+          target/release/examples/dlint benchmarks/benchmarks.ts
 
       - name: Benchmarks
         if: contains(matrix.os, 'ubuntu')
         run: deno run -A --quiet benchmarks/benchmarks.ts
-
-      - name: Pre-release (linux)
-        if: |
-          contains(matrix.os, 'ubuntu')
-        run: |
-          cd target/release
-          zip -r deno_lint-x86_64-unknown-linux-gnu.zip deno_lint
-
-      - name: Pre-release (mac)
-        if: |
-          contains(matrix.os, 'macOS')
-        run: |
-          cd target/release
-          zip -r deno_lint-x86_64-apple-darwin.zip deno_lint
-
-      - name: Pre-release (windows)
-        if: |
-          contains(matrix.os, 'windows')
-        run: |
-          Compress-Archive -CompressionLevel Optimal -Force -Path target/release/deno_lint.exe -DestinationPath target/release/deno_lint-x86_64-pc-windows-msvc.zip
-
-      - name: Release
-        uses: softprops/action-gh-release@v1
-        if: |
-          github.repository == 'denoland/deno_lint' &&
-          startsWith(github.ref, 'refs/tags/')
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        with:
-          files: |
-            target/release/deno_lint-x86_64-pc-windows-msvc.zip
-            target/release/deno_lint-x86_64-unknown-linux-gnu.zip
-            target/release/deno_lint-x86_64-apple-darwin.zip
-          draft: true
 
       - name: Publish
         if: |
