+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "deno_lint"
-version = "0.1.6"
+version = "0.1.7"
 edition = "2018"
 description = "lint for deno"
 authors = ["the Deno authors"]
@@ -23,7 +23,7 @@ swc_ecma_ast = { version = "=0.20.0" }
 swc_ecma_parser = { version = "=0.24.5" }
 swc_ecma_visit = { version = "=0.5.1" }
 regex = "1.3.6"
-termcolor = "1.1.0"
 
 [dev-dependencies]
 clap = { version = "2.33.1" }
+termcolor = "1.1.0"
