+++ examples/dlint/main.rs
@@ -1,6 +1,49 @@
 // Copyright 2020 the Deno authors. All rights reserved. MIT license.
 use clap::App;
 use clap::Arg;
+use deno_lint::diagnostic::LintDiagnostic;
+use deno_lint::linter::Linter;
+use deno_lint::rules::get_all_rules;
+use std::fmt;
+use std::io::Write;
+use termcolor::Color::{Ansi256, Red};
+use termcolor::{Ansi, ColorSpec, WriteColor};
+
+#[cfg(windows)]
+use termcolor::{BufferWriter, ColorChoice};
+
+#[allow(unused)]
+#[cfg(windows)]
+fn enable_ansi() {
+  BufferWriter::stdout(ColorChoice::AlwaysAnsi);
+}
+
+fn gray(s: String) -> impl fmt::Display {
+  let mut style_spec = ColorSpec::new();
+  style_spec.set_fg(Some(Ansi256(8)));
+  style(&s, style_spec)
+}
+
+fn red(s: String) -> impl fmt::Display {
+  let mut style_spec = ColorSpec::new();
+  style_spec.set_fg(Some(Red));
+  style(&s, style_spec)
+}
+
+fn cyan(s: String) -> impl fmt::Display {
+  let mut style_spec = ColorSpec::new();
+  style_spec.set_fg(Some(Ansi256(14)));
+  style(&s, style_spec)
+}
+
+fn style(s: &str, colorspec: ColorSpec) -> impl fmt::Display {
+  let mut v = Vec::new();
+  let mut ansi_writer = Ansi::new(&mut v);
+  ansi_writer.set_color(&colorspec).unwrap();
+  ansi_writer.write_all(s.as_bytes()).unwrap();
+  ansi_writer.reset().unwrap();
+  String::from_utf8_lossy(&v).into_owned()
+}
 
 fn create_cli_app<'a, 'b>(rule_list: &'b str) -> App<'a, 'b> {
   App::new("dlint").after_help(rule_list).arg(
@@ -11,13 +54,58 @@ fn create_cli_app<'a, 'b>(rule_list: &'b str) -> App<'a, 'b> {
   )
 }
 
-fn main() {
-  use deno_lint::colors;
-  use deno_lint::linter::Linter;
-  use deno_lint::rules::get_all_rules;
+pub fn format_diagnostic(diagnostic: &LintDiagnostic) -> String {
+  let pretty_error = format!(
+    "({}) {}",
+    gray(diagnostic.code.to_string()),
+    diagnostic.message
+  );
 
+  let file_name = &diagnostic.location.filename;
+  let location = if file_name.contains('/')
+    || file_name.contains('\\')
+    || file_name.starts_with("./")
+  {
+    file_name.to_string()
+  } else {
+    format!("./{}", file_name)
+  };
+
+  let line_str_len = diagnostic.location.line.to_string().len();
+  let pretty_location = cyan(format!(
+    "{}--> {}:{}:{}",
+    " ".repeat(line_str_len),
+    location,
+    diagnostic.location.line,
+    diagnostic.location.col
+  ))
+  .to_string();
+
+  let dummy = format!("{} |", " ".repeat(line_str_len));
+  let pretty_line_src =
+    format!("{} | {}", diagnostic.location.line, diagnostic.line_src);
+  let red_glyphs = format!(
+    "{} | {}{}",
+    " ".repeat(line_str_len),
+    " ".repeat(diagnostic.location.col),
+    red("^".repeat(diagnostic.snippet_length))
+  );
+
+  let lines = vec![
+    pretty_error,
+    pretty_location,
+    dummy.clone(),
+    pretty_line_src,
+    red_glyphs,
+    dummy,
+  ];
+
+  lines.join("\n")
+}
+
+fn main() {
   #[cfg(windows)]
-  colors::enable_ansi();
+  enable_ansi();
 
   let rules = get_all_rules();
 
@@ -51,7 +139,7 @@ fn main() {
 
     error_counts += file_diagnostics.len();
     for d in file_diagnostics.iter() {
-      eprintln!("{}", d.to_pretty_string());
+      eprintln!("{}", format_diagnostic(d));
     }
   }
