--- src/colors.rs
@@ -1,54 +1,0 @@
-// Copyright 2020 the Deno authors. All rights reserved. MIT license.
-
-use std::env;
-use std::fmt;
-use std::io::Write;
-use termcolor::Color::{Ansi256, Red};
-use termcolor::{Ansi, ColorSpec, WriteColor};
-
-lazy_static! {
-  static ref NO_COLOR: bool = { env::var_os("NO_COLOR").is_some() };
-}
-
-pub fn use_color() -> bool {
-  !(*NO_COLOR)
-}
-
-#[cfg(windows)]
-use termcolor::{BufferWriter, ColorChoice};
-
-#[allow(unused)]
-#[cfg(windows)]
-pub fn enable_ansi() {
-  BufferWriter::stdout(ColorChoice::AlwaysAnsi);
-}
-
-pub fn gray(s: String) -> impl fmt::Display {
-  let mut style_spec = ColorSpec::new();
-  style_spec.set_fg(Some(Ansi256(8)));
-  style(&s, style_spec)
-}
-
-pub fn red(s: String) -> impl fmt::Display {
-  let mut style_spec = ColorSpec::new();
-  style_spec.set_fg(Some(Red));
-  style(&s, style_spec)
-}
-
-pub fn cyan(s: String) -> impl fmt::Display {
-  let mut style_spec = ColorSpec::new();
-  style_spec.set_fg(Some(Ansi256(14)));
-  style(&s, style_spec)
-}
-
-fn style(s: &str, colorspec: ColorSpec) -> impl fmt::Display {
-  if !use_color() {
-    return String::from(s);
-  }
-  let mut v = Vec::new();
-  let mut ansi_writer = Ansi::new(&mut v);
-  ansi_writer.set_color(&colorspec).unwrap();
-  ansi_writer.write_all(s.as_bytes()).unwrap();
-  ansi_writer.reset().unwrap();
-  String::from_utf8_lossy(&v).into_owned()
-}
